Scale PDFs correctly (#52)
* graphs: remove groupid to reduce noise

If you need to know exactly what dependency is used, you would look at the pom.xml file anyway.

* scale PDF to correct size

* Region: add with* methods

* images: update mockito use in test

+++ spi/src/main/java/net/kemitix/clover/spi/PdfWidth.java
@@ -3,5 +3,5 @@ package net.kemitix.clover.spi;
 import net.kemitix.properties.typed.TypedProperty;
 
 public interface PdfWidth
-        extends TypedProperty<Integer> {
+        extends TypedProperty<Float> {
 }
